+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   This is a breaking change mostly because items with `Snippet` kind are no longer filtered out by default.
 - FEATURE: add scrolling in info and signature window. By default can be done with `<C-f>` / `<C-b>` when target window is shown. Can be configured via `mappings.scroll_down` and `mappings.scroll_up` config options.
 - FEATURE: respect `isIncomplete` in LSP completion response and immediately force new completion request on the next key press.
+- FEATURE: add support for context in 'textDocument/completion' request.
 - FEATURE: both info and signature help windows now use tree-sitter highlighting:
     - Info window uses "markdown" parser (works best on Neovim>=0.10 as its parser is built-in). Special markdown characters are concealed (i.e. hidden) which might result into seemingly unnecessary whitespace as dimensions are computed not accounting for that.
     - Signature help uses same parser as in current filetype.